refactor: Separate retry logic into RetryFailedCommandHandler and update related tests
