+++ ActionProcessor/Program.cs
@@ -53,6 +53,7 @@ builder.Services.AddScoped<IEventRepository, EventRepository>();
 
 // Application Services
 builder.Services.AddScoped<FileCommandHandler>();
+builder.Services.AddScoped<RetryFailedCommandHandler>();
 builder.Services.AddScoped<FileQueryHandler>();
 
 // Validators
